* Minor changes:   - Removed parameters passed to createSipMessageXXXX(...) that were not     used (string sipDomain?!)   - Removed some commented out code   - Fixed some warnings for things that are ok (looked through the output     after compiling with a bit more strict settings than usual)
git-svn-id: svn://svn.minisip.org/minisip/trunk@2097 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/minisip/gui/textui/MinisipTextUI.cxx
@@ -271,7 +271,7 @@ void MinisipTextUI::handleCommand(const CommandString &cmd){
 		string confid="";
 		string users=cmd.getParam3();
 		int i=0;	
-		while (users[i]!=';'&&users.length()!=0 &&!(i>(users.length()-1))){
+		while (users[i]!=';'&&users.length()!=0 &&!(i>((int)users.length()-1))){
 			confid=confid+users[i];
 			i++;
 		}
